Split field functions into separate module for clarity
