+++ dev/lfdev/dev_vistools_2dplatehole.py
@@ -15,8 +15,7 @@ import pyvale
 
 def main() -> None:
     data_path = Path('src/data/case17_out.e')
-    data_reader = mh.ExodusReader(data_path)
-    sim_data = data_reader.read_all_sim_data()
+sim_data = mh.ExodusReader(data_path).read_all_sim_data()
     # Scale to mm to make 3D visualisation scaling easier
     sim_data.coords = sim_data.coords*1000.0 # type: ignore
