+++ dev/lfdev/dev_vistools_3dmonoblock.py
@@ -18,8 +18,7 @@ def main() -> None:
     """
     # Use mooseherder to read the exodus and get a SimData object
     data_path = Path('src/data/case16_out.e')
-    data_reader = mh.ExodusReader(data_path)
-    sim_data = data_reader.read_all_sim_data()
+sim_data = mh.ExodusReader(data_path).read_all_sim_data()
     field_name = 'temperature'
     # Scale to mm to make 3D visualisation scaling easier
     sim_data.coords = sim_data.coords*1000.0 # type: ignore
