+++ dev/lfdev/figuregen/ex3_displacement2d_basicvis.py
@@ -16,8 +16,7 @@ def main() -> None:
     """pyvale example:
     """
     data_path = Path('simcases/case17/case17_out.e')
-    data_reader = mh.ExodusReader(data_path)
-    sim_data = data_reader.read_all_sim_data()
+sim_data = mh.ExodusReader(data_path).read_all_sim_data()
     # Scale to mm to make 3D visualisation scaling easier
     sim_data.coords = sim_data.coords*1000.0 # type: ignore
