+++ src/pyvale/core/fieldvector.py
@@ -10,9 +10,9 @@ import pyvista as pv
 from scipy.spatial.transform import Rotation
 import mooseherder as mh
 
-from pyvale.core.field import (IField,
-                          conv_simdata_to_pyvista,
-                          sample_pyvista)
+from pyvale.core.field import IField
+from pyvale.core.fieldconverter import conv_simdata_to_pyvista
+from pyvale.core.fieldsampler import sample_pyvista_grid
 from pyvale.core.fieldtransform import (transform_vector_2d,
                                    transform_vector_2d_batch,
                                    transform_vector_3d,
@@ -68,7 +68,7 @@ class FieldVector(IField):
                     angles: tuple[Rotation,...] | None = None,
                     ) -> np.ndarray:
 
-        field_data = sample_pyvista(self._components,
+        field_data = sample_pyvista_grid(self._components,
                                 self._pyvista_grid,
                                 self._sim_data.time,
                                 points,
