+++ src/pyvale/core/visualtools.py
@@ -12,10 +12,10 @@ import numpy as np
 #NOTE: causes output to console to be suppressed unfortunately
 import pyvista as pv
 from pyvale.core.visualopts import (VisOptsSimSensors,
-                               VisOptsImageSave,
-                               EImageType,
-                               VisOptsAnimation,
-                               EAnimationType)
+                                    VisOptsImageSave,
+                                    EImageType,
+                                    VisOptsAnimation,
+                                    EAnimationType)
 
 
 def create_pv_plotter(vis_opts: VisOptsSimSensors) -> pv.Plotter:
