+++ src/pyvale/examples/ex2_2_thermal3d.py
@@ -16,8 +16,7 @@ def main() -> None:
     ----------------------------------------------------------------------------
     """
     data_path = pyvale.DataSet.thermal_3d_output_path()
-    data_reader = mh.ExodusReader(data_path)
-    sim_data = data_reader.read_all_sim_data()
+    sim_data = mh.ExodusReader(data_path).read_all_sim_data()
     field_name = 'temperature'
 
     # Scale to mm to make 3D visualisation scaling easier
