+++ src/pyvale/examples/ex4_2_strain2d.py
@@ -17,8 +17,7 @@ def main() -> None:
     ----------------------------------------------------------------------------
     """
     data_path = pyvale.DataSet.mechanical_2d_output_path()
-    data_reader = mh.ExodusReader(data_path)
-    sim_data = data_reader.read_all_sim_data()
+    sim_data = mh.ExodusReader(data_path).read_all_sim_data()
     # Scale to mm to make 3D visualisation scaling easier
     sim_data.coords = sim_data.coords*1000.0 # type: ignore
